Notify the loading failed if there is no network
When we get to the fetchArticles and we find that there is no network,
we should treat it like a failure and call notifyLoadMoreFailed on the
EndlessRecyclerViewScrollListener. This case is actually quite rare
because SearchActivity has a BroadcaseReceiver that disables the call of
onLoadMore (which calls fetchArticles) when the network goes down.